refactor: move SubPathsFor upper
